lesson-02-data-engineering/homework/requirements.txt
# Document parsing (replaces PyPDF2, python-docx, beautifulsoup4, pytesseract)
unstructured[pdf,docx,xlsx]>=0.16.0

# Chunking (replaces hand-rolled chunker)
langchain-text-splitters>=0.3.0

# Encoding detection (replaces hand-rolled EncodingNormalizer)
charset-normalizer>=3.0.0

# File type detection via magic bytes (replaces hand-rolled MAGIC_SIGNATURES)
filetype>=1.2.0

# Retry with backoff (replaces hand-rolled retry_with_backoff)
tenacity>=9.0.0

# Streaming / watching
watchdog>=4.0.0

# Config
PyYAML>=6.0.0

# Sample generation (still needed for generate_samples.py)
openpyxl>=3.1.0
python-docx>=1.1.0

# Testing
pytest>=8.0.0
